fix(temporal): fix CI lint + test failures
- Lint (max-lines): github-webhook.ts exceeded the 500-line cap (533).
  Extract the Zod webhook schemas, RELEVANT_ACTIONS, and the author-trust
  check into a new github-webhook-schema.ts (475 lines remaining).
- Test (ingest-dismissals): the PR added a security gate requiring the 👎
  reaction's user.login to equal the repo owner. Update the two thumbs-down
  fixtures to include `user: { login: "o" }` so they reflect the new contract.
- Test (mock leak): agent-task.test.ts's process-wide mock.module replaced
  the entire agent-task-command module, dropping reportOnlyPrompt and breaking
  alert-remediation-command.test.ts. Spread the real module's exports back in.
- Test (schedules): categorize alertRemediationSweepWorkflow in
  WORKFLOWS_WITHOUT_LONG_SLEEPS (matches main; lost in an earlier merge).
