+++ packages/temporal/src/event-bridge/github-webhook.ts
@@ -1,7 +1,6 @@
 import { Hono } from "hono";
 import { verify } from "@octokit/webhooks-methods";
 import { Octokit } from "octokit";
-import { z } from "zod/v4";
 import * as Sentry from "@sentry/bun";
 import type { Client } from "@temporalio/client";
 import { WorkflowIdReusePolicy } from "@temporalio/common";
@@ -41,67 +40,14 @@ import {
   createGitHubAppInstallationToken,
   type GitHubAppTokenResult,
 } from "#lib/github-app-token.ts";
+import {
+  PullRequestEventSchema,
+  RELEVANT_ACTIONS,
+  disallowedAuthorReason,
+} from "./github-webhook-schema.ts";
 
 const COMPONENT = "pr-webhook";
 const DEFAULT_PORT = 9466;
-
-const RELEVANT_ACTIONS = new Set([
-  "opened",
-  "synchronize",
-  "reopened",
-  "ready_for_review",
-]);
-
-// Security: PR automation (the review/summary pipelines — whose verify stage
-// checks out and executes PR-head code) must only run for the trusted
-// repository owner. The repo is public, so without this gate any external
-// fork PR's title/diff/code would flow into the agents and the verifier.
-const ALLOWED_PR_AUTHOR = "shepherdjerred";
-
-// Returns a skip reason (for metrics/logs) when a PR's author is not the
-// trusted owner — bots and any non-owner account are skipped — or null to
-// proceed.
-function disallowedAuthorReason(user: {
-  readonly login: string;
-  readonly type: string;
-}): string | null {
-  if (user.type === "Bot") return "bot-author";
-  if (user.login !== ALLOWED_PR_AUTHOR) return "untrusted-author";
-  return null;
-}
-
-const PrUserSchema = z.object({
-  login: z.string(),
-  type: z.string(),
-});
-
-const PrRefSchema = z.object({
-  ref: z.string(),
-  sha: z.string(),
-});
-
-const PrSchema = z.object({
-  number: z.number().int().positive(),
-  draft: z.boolean().optional(),
-  merged: z.boolean().optional(),
-  title: z.string(),
-  base: PrRefSchema,
-  head: PrRefSchema,
-  user: PrUserSchema,
-});
-
-const RepoOwnerSchema = z.object({ login: z.string() });
-
-const RepoSchema = z.object({
-  name: z.string(),
-  owner: RepoOwnerSchema,
-});
-
-const PullRequestEventSchema = z.object({
-  action: z.string(),
-  pull_request: PrSchema,
-  repository: RepoSchema,
-});
 
 export type WebhookHandle = {
   port: number;
